implement crossover

+++ src/headers/util.h
@@ -14,7 +14,7 @@ class Util {
     // public constructor
     // public methods
     // returns a random number between 0 and 1
-    static float random() {
+    static double random() {
         static std::random_device rd;
         static std::mt19937 gen(rd());
         static std::uniform_real_distribution<> dis(0, 1);
